Refactor sampler in independent module
